+++ work/resources/ts/containers/organisms/SignIn.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+/**
+ * Please import if necessary
+*/
+// import { StylesContext } from '../../contexts/childContexts/StylesContext';
+// import {  } from '@material-ui/core';
+// import {  } from '@material-ui/icons';
+
+type Props = {
+  // Defines the type of props.
+
+}
+
+export const Component: React.FC<Props> = (props: Props) => {
+  return (
+    <>
+      {/* Please describe tsx */}
+
+    </>
+  );
+}
