+++ work/resources/ts/containers/pages/SignInPage.tsx
@@ -1,0 +1,26 @@
+import React, { useContext } from 'react';
+import { Header } from '../organisms/Header';
+import { StylesContext } from '../../contexts/childContexts/StylesContext';
+import {
+  Container,
+  CssBaseline,
+  Grid,
+  // useMediaQuery,
+} from '@material-ui/core';
+
+export const SignInPage: React.FC = () => {
+  const { useStyles } = useContext<any>(StylesContext);
+  const classes = useStyles();
+
+  return (
+    <>
+      <CssBaseline />
+      <Header
+        title='makala'
+      />
+      <Container maxWidth='xl' className={classes.main_container}>
+        <p>aaa</p>
+      </Container>
+    </>
+  )
+}
